requirements.txt
scikit-learn
numpy
pandas
matplotlib
seaborn
fastapi
uvicorn
pydantic
ipykernel
simpy
supabase
python-dotenv
dill
pyspark
pyspark[sql]
pyspark[pandas_on_spark]
sqlmodel
psycopg2